refactor: purchase error code 패키지 수정
- 잘못 위치한 purchase error code 패키지 수정함
