+++ backend/build.gradle
@@ -154,6 +154,7 @@ subprojects {
                         '**/global/docs/**', // exclude docs package
                         '**/global/exception/**', // exclude exception package
                         '**/global/utils/**', // exclude utils package
+                        '**/domain/purchase/exception/**',
                         '**/domain/purchase/repository/**', // todo 테스트 설정 에러 픽스 후 삭제
                         '**/global/log/**', // exclude utils package
                 ])
